+++ src/parser/minc/minc_internal.h
@@ -194,7 +194,7 @@ void yyerror(char *msg);
 struct symbol *install(char *name, ScopeType scope);
 struct symbol *lookup(char *name);
 char *strsave(char *str);
-char *emalloc(int nbytes);
+char *emalloc(long nbytes);
 void efree(void *mem);
 void clear_elem(MincListElem *);
 void unref_value_list(MincValue *);
